requirements.txt
fastapi
uvicorn[standard]
python-multipart

nfstream
numpy
pandas
PyYAML
requests

SQLAlchemy
psycopg2-binary

streamlit
streamlit-autorefresh

tqdm
